switch to using useSelector for getting data from the store
